requirements.txt
Pillow
pytest
coverage
black[d]